# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartRest AIoT API",
        "description": "RESTful API for smart mattress monitoring system supporting patients, doctors, customers, and admins.",
        "contact": {
            "name": "SmartRest Support",
            "email": "support@smartrest.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "SmartRest API Server"
        }
    ],
    "paths": {
        "/analytics/sleep-report": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get sleep report",
                "description": "Retrieve AI-generated sleep report including sleep stages, posture map, and overall sleep score",
                "operationId": "getSleepReport",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date for the report (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID (required for doctors/admins, not for patients)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sleep report data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "patient_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "sleep_duration": {
                                            "properties": {
                                                "total_hours": {
                                                    "type": "number",
                                                    "example": 7.5
                                                },
                                                "minutes": {
                                                    "type": "integer",
                                                    "example": 450
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "sleep_stages": {
                                            "properties": {
                                                "awake": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "light": {
                                                    "type": "integer",
                                                    "example": 210
                                                },
                                                "deep": {
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "rem": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "sleep_score": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "posture_changes": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "breathing_events": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "avg_heart_rate": {
                                            "type": "number",
                                            "example": 68
                                        },
                                        "notes": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - No access to this patient's data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Patient profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}